requirements.txt
numpy
scipy
pandas
openpyxl
pycryptodome
pillow
